refactor: split parser utilities
