+++ src/logic/workers/mapping/otlpMapper.ts
@@ -1,177 +1,23 @@
 /**
  * Convert raw OTLP JSON payload into the internal ParsedSnapshot graph.
  *
- * {@markdown
- * ### Instrument Type Mapping
- * | OTLP field present | instrumentType | temporality rule | isMonotonic |
- * |--------------------|----------------|------------------|-------------|
- * | gauge.dataPoints   | Gauge          | Unspecified      | undefined   |
- * | sum.dataPoints     | Sum            | Map enum AGGREGATION_TEMPORALITY_* → string | copy |
- * | histogram.dataPoints | Histogram    | same enum mapping | undefined |
- * | (future) summary.dataPoints | Summary | Unspecified | n/a |
- * }
- *
- * {@markdown
- * ### Detailed Algorithm
- * 1. Create ParsedSnapshot root with id, fileName and current timestamp.
- * 2. For each `resourceMetrics` entry:
- *    - Map resource attributes.
- *    - For each `scopeMetrics` entry:
- *      - Capture scope name/version/attributes.
- *      - For each metric:
- *        - Derive `MetricDefinition` and locate its data points.
- *        - For every point:
- *          - Merge resource and metric attributes and encode a SeriesKey.
- *          - Map the raw point to a typed ParsedPoint.
- *          - Append the point under the SeriesKey in a Map.
- *        - Push ParsedMetricData with definition and series map.
- *      - Push ParsedScopeData to resource node.
- *    - Push ParsedResourceData to snapshot.
- * 3. Return the populated snapshot.
- * }
- *
- * {@markdown
- * ### Performance Notes
- * | Dataset                 | Size  | Mapping time (worker) |
- * |-------------------------|-------|-----------------------|
- * | 100 k series, 1 pt      | 25 MB | ~110 ms               |
- * | 10 k series, 10 pts     | 18 MB | ~85 ms                |
- * }
- * No intermediate arrays beyond what GC needs; attribute maps are reused where safe.
- *
- * {@markdown
- * ### Tests
- * - Minimal gauge fixture parses to a single series.
- * - Histogram with exemplars preserves exemplar count.
- * - Duplicate attribute order results in a single series (order-insensitivity).
- * - Unknown metric kind throws an error.
- * - 25 MB fixture maps in under 150 ms.
- * }
+ * See the original file for detailed algorithm and performance notes.
  */
-
-import type { RawOtlpExportMetricsServiceRequest, RawOtlpMetric, RawOtlpNumberDataPoint, RawOtlpHistogramDataPoint, RawOtlpKeyValue } from '@intellimetric/contracts/rawOtlpTypes';
-import type { ParsedSnapshot, ParsedResourceData, ParsedScopeData, ParsedMetricData, ParsedSeriesData, ParsedPoint, MetricDefinition, AttrMap } from '@intellimetric/contracts/types';
+import type {
+  RawOtlpExportMetricsServiceRequest,
+} from '@intellimetric/contracts/rawOtlpTypes';
+import type {
+  ParsedSnapshot,
+  ParsedResourceData,
+  ParsedScopeData,
+  ParsedMetricData,
+  ParsedSeriesData,
+} from '@intellimetric/contracts/types';
 import { encodeSeriesKey } from '../utils/seriesKeyEncoder';
-import { extractExemplars } from '../utils/exemplarExtractor';
+import { mapAttrs, mapPoint } from './mappingUtils';
+import { deriveMetric } from './metricDerivation';
 
-function mapAttrs(raw?: RawOtlpKeyValue[] | null): AttrMap {
-  const out: AttrMap = {};
-  if (!raw) return out;
-  for (const kv of raw) {
-    if (!kv || !kv.value) continue;
-    const v = kv.value as any;
-    if (v.stringValue !== undefined) out[kv.key] = v.stringValue;
-    else if (v.doubleValue !== undefined) out[kv.key] = v.doubleValue;
-    else if (v.intValue !== undefined) out[kv.key] = Number(v.intValue);
-    else if (v.boolValue !== undefined) out[kv.key] = v.boolValue;
-  }
-  return out;
-}
-
-function mapTemporality(t: number | undefined): 'Delta' | 'Cumulative' | 'Unspecified' {
-  switch (t) {
-    case 1:
-      return 'Delta';
-    case 2:
-      return 'Cumulative';
-    default:
-      return 'Unspecified';
-  }
-}
-
-interface MetricInfo {
-  definition: MetricDefinition;
-  points: Array<RawOtlpNumberDataPoint | RawOtlpHistogramDataPoint>;
-}
-
-function deriveMetric(m: RawOtlpMetric): MetricInfo {
-  if (m.gauge?.dataPoints) {
-    return {
-      definition: {
-        name: m.name,
-        description: m.description,
-        unit: m.unit,
-        instrumentType: 'Gauge',
-        temporality: 'Unspecified',
-      },
-      points: m.gauge.dataPoints as RawOtlpNumberDataPoint[],
-    };
-  }
-
-  if (m.sum?.dataPoints) {
-    return {
-      definition: {
-        name: m.name,
-        description: m.description,
-        unit: m.unit,
-        instrumentType: 'Sum',
-        temporality: mapTemporality(m.sum.aggregationTemporality),
-        isMonotonic: m.sum.isMonotonic,
-      },
-      points: m.sum.dataPoints as RawOtlpNumberDataPoint[],
-    };
-  }
-
-  if (m.histogram?.dataPoints) {
-    return {
-      definition: {
-        name: m.name,
-        description: m.description,
-        unit: m.unit,
-        instrumentType: 'Histogram',
-        temporality: mapTemporality(m.histogram.aggregationTemporality),
-      },
-      points: m.histogram.dataPoints as RawOtlpHistogramDataPoint[],
-    };
-  }
-
-  return {
-    definition: {
-      name: m.name,
-      description: m.description,
-      unit: m.unit,
-      instrumentType: 'Unknown',
-    },
-    points: [],
-  };
-}
-
-function mapNumberPoint(raw: RawOtlpNumberDataPoint): ParsedPoint {
-  return {
-    timestampUnixNano: Number(raw.timeUnixNano),
-    startTimeUnixNano: raw.startTimeUnixNano ? Number(raw.startTimeUnixNano) : undefined,
-    value: raw.asDouble ?? Number(raw.asInt ?? 0),
-    attributes: mapAttrs(raw.attributes),
-    exemplars: extractExemplars(raw.exemplars),
-  };
-}
-
-function mapHistogramPoint(raw: RawOtlpHistogramDataPoint): ParsedPoint {
-  return {
-    timestampUnixNano: Number(raw.timeUnixNano),
-    startTimeUnixNano: raw.startTimeUnixNano ? Number(raw.startTimeUnixNano) : undefined,
-    count: Number(raw.count),
-    sum: raw.sum !== undefined ? Number(raw.sum) : undefined,
-    bucketCounts: (raw.bucketCounts || []).map(Number),
-    explicitBounds: (raw.explicitBounds || []).map(Number),
-    min: raw.min !== undefined ? Number(raw.min) : undefined,
-    max: raw.max !== undefined ? Number(raw.max) : undefined,
-    attributes: mapAttrs(raw.attributes),
-    exemplars: extractExemplars(raw.exemplars),
-  };
-}
-
-function mapPoint(raw: RawOtlpNumberDataPoint | RawOtlpHistogramDataPoint, type: MetricDefinition['instrumentType']): ParsedPoint {
-  if (type === 'Gauge' || type === 'Sum') return mapNumberPoint(raw as RawOtlpNumberDataPoint);
-  if (type === 'Histogram') return mapHistogramPoint(raw as RawOtlpHistogramDataPoint);
-  throw new Error('Unsupported OTLP metric shape: ' + type);
-}
-
-/**
- * Transform a parsed OTLP JSON object into ParsedSnapshot.
- *
- * @throws Error on unsupported metric/data-point shape.
- */
+/** Transform a parsed OTLP object into a ParsedSnapshot structure. */
 export function mapToParsedSnapshot(
   raw: RawOtlpExportMetricsServiceRequest,
   snapshotId: string,
@@ -201,14 +47,21 @@ export function mapToParsedSnapshot(
       const metrics = scope.metrics || [];
       for (const m of metrics) {
         const metricInfo = deriveMetric(m);
-        if (metricInfo.definition.instrumentType === 'Unknown' || !metricInfo.points) {
+        if (
+          metricInfo.definition.instrumentType === 'Unknown' ||
+          !metricInfo.points
+        ) {
           throw new Error('Unsupported OTLP metric shape: ' + m.name);
         }
 
         const seriesMap: Map<string, ParsedSeriesData> = new Map();
         for (const pt of metricInfo.points) {
           const metricAttrs = mapAttrs(pt.attributes);
-          const seriesKey = encodeSeriesKey(metricInfo.definition.name, rAttrs, metricAttrs);
+          const seriesKey = encodeSeriesKey(
+            metricInfo.definition.name,
+            rAttrs,
+            metricAttrs,
+          );
           const parsedPt = mapPoint(pt, metricInfo.definition.instrumentType);
 
           if (!seriesMap.has(seriesKey)) {
@@ -223,7 +76,10 @@ export function mapToParsedSnapshot(
           seriesMap.get(seriesKey)!.points.push(parsedPt);
         }
 
-        sNode.metrics.push({ definition: metricInfo.definition, seriesData: seriesMap } as ParsedMetricData);
+        sNode.metrics.push({
+          definition: metricInfo.definition,
+          seriesData: seriesMap,
+        } as ParsedMetricData);
       }
 
       rNode.scopes.push(sNode);
@@ -234,4 +90,3 @@ export function mapToParsedSnapshot(
 
   return snapshot;
 }
-
